requirements.txt
annotated-types>=0.7.0
attrs>=24.2.0
beautifulsoup4>=4.13.3
black>=24.10.0
blinker>=1.9.0
cachetools>=5.5.0
certifi>=2024.8.30
cffi>=1.17.1
charset-normalizer>=3.3.2
click>=8.1.7
colorama>=0.4.6
contourpy>=1.3.0
cycler>=0.12.1
dacite>=1.8.1
db-dtypes>=1.3.0
dill>=0.3.9
et_xmlfile>=2.0.0
excel-util>=0.0.4
exceptiongroup>=1.2.2
Flask>=3.1.0
flask-cors>=5.0.1
fonttools>=4.54.1
google-api-core>=2.20.0
google-api-python-client>=2.147.0
google-auth>=2.35.0
google-auth-httplib2>=0.2.0
google-cloud-bigquery>=3.26.0
google-cloud-bigquery-datatransfer>=3.17.1
google-cloud-bigquery-storage>=2.13.0
google-cloud-core>=2.4.1
google-cloud-storage>=2.3.0
google-crc32c>=1.6.0
google-oauth2-tool>=0.0.3
google-resumable-media>=2.7.2
googleapis-common-protos>=1.65.0
grpcio>=1.66.2
grpcio-status>=1.66.2
h11>=0.14.0
htmlmin>=0.1.12
httplib2>=0.22.0
idna>=3.10
ImageHash>=4.3.1
importlib_metadata>=8.5.0
importlib_resources>=6.4.5
iniconfig>=2.0.0
itsdangerous>=2.2.0
Jinja2>=3.1.4
joblib>=1.4.2
kiwisolver>=1.4.7
llvmlite>=0.43.0
markdown-it-py>=3.0.0
MarkupSafe>=3.0.2
matplotlib>=3.9.2
mdurl>=0.1.2
mouse>=0.7.1
MouseInfo>=0.1.3
multimethod>=1.12
multiprocess>=0.70.17
mypy-extensions>=1.0.0
networkx>=3.2.1
numba>=0.60.0
numpy>=2.0.2
oauth2client>=4.1.3
openpyxl>=3.1.5
outcome>=1.3.0.post0
packaging>=24.1
pandas>=2.2.3
pandas-util>=0.0.1
pathspec>=0.12.1
patsy>=1.0.1
phik>=0.12.4
pillow>=11.0.0
platformdirs>=4.3.6
pluggy>=1.5.0
proto-plus>=1.24.0
protobuf>=5.28.2
pyarrow>=17.0.0
pyasn1>=0.6.1
pyasn1_modules>=0.4.1
PyAutoGUI>=0.9.54
pycparser>=2.22
pydantic>=2.9.2
pydantic_core>=2.23.4
PyGetWindow>=0.0.9
Pygments>=2.18.0
PyMsgBox>=1.0.9
pyparsing>=3.1.4
pyperclip>=1.9.0
PyRect>=0.2.0
PyScreeze>=1.0.1
PySocks>=1.7.1
pytest>=8.3.5
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
pytweening>=1.2.0
pytz>=2024.2
PyWavelets>=1.6.0
PyYAML>=6.0.2
requests>=2.32.3
rich>=13.9.4
rsa>=4.9
scipy>=1.13.1
seaborn>=0.13.2
selenium>=4.28.1
six>=1.16.0
sniffio>=1.3.1
sortedcontainers>=2.4.0
soupsieve>=2.6
statsmodels>=0.14.4
tomli>=2.0.2
tqdm>=4.67.0
trio>=0.29.0
trio-websocket>=0.12.0
typeguard>=4.4.1
typing_extensions>=4.12.2
tzdata>=2024.2
uritemplate>=4.1.1
urllib3>=2.2.3
visions>=0.7.6
webdriver-manager>=4.0.2
websocket-client>=1.8.0
Werkzeug>=3.1.3
wordcloud>=1.9.4
wsproto>=1.2.0
ydata-profiling>=4.12.0
zipp>=3.21.0
fsspec>=2025.7.0
gcsfs>=2025.7.0